Extract Feed Loader Stub into a shared scope to remove duplication
